Extract common utilities for both F2 and gmos imaging services
